+++ bindings/java/native/src/types/info_response.rs
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use getset::{CopyGetters, Getters};
-use iota::{
-    bee_rest_api::types::responses::InfoResponse as RustInfoResponse, client::NodeInfoWrapper as RustNodeInfoWrapper,
+use iota_client::{
+    bee_rest_api::types::responses::InfoResponse as RustInfoResponse, NodeInfoWrapper as RustNodeInfoWrapper,
 };
 
 #[derive(PartialEq, Getters, CopyGetters)]