+++ bindings/nodejs/lib/index.d.ts
@@ -294,7 +294,7 @@ export declare class AccountManager {
   removeAccount(accountId: string | number): void
   syncAccounts(options?: SyncOptions): Promise<SyncedAccount[]>
   internalTransfer(fromAccount: Account, toAccount: Account, amount: number): Promise<Message>
-  backup(destination: string): string
+  backup(destination: string, password: string): string
   importAccounts(source: string, password: string): void
   isLatestAddressUnused(): Promise<boolean>
   setClientOptions(options: ClientOptions): void
@@ -321,7 +321,8 @@ export declare type Event = 'ErrorThrown' |
   'ConfirmationStateChange' |
   'Reattachment' |
   'Broadcast' |
-  'TransferProgress'
+  'TransferProgress' |
+  'MigrationProgress'
 
 export interface LoggerOutput {
   name?: string