+++ bindings/nodejs/native/src/classes/account/tasks/node_info.rs
@@ -1,7 +1,7 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use iota::client::NodeInfoWrapper;
+use iota_client::NodeInfoWrapper;
 use iota_wallet::Error;
 use neon::prelude::*;