+++ src/monitor.rs
@@ -9,7 +9,7 @@ use crate::{
     message::{Message, MessagePayload, TransactionEssence, TransactionInput, TransactionOutput},
 };
 
-use iota::{
+use iota_client::{
     bee_rest_api::types::{dtos::OutputDto, responses::OutputResponse},
     Topic, TopicEvent,
 };
@@ -66,7 +66,9 @@ async fn subscribe_to_topics<C: Fn(&TopicEvent) + Send + Sync + 'static>(
         tokio::spawn(async move {
             let client = crate::client::get_client(&client_options).await?;
             let mut client = client.write().await;
-            client.subscriber().with_topics(topics).subscribe(handler).await?;
+            if let Err(err) = client.subscriber().with_topics(topics).subscribe(handler).await {
+                log::debug!("[MQTT] subscribe error: {:?}", err);
+            }
             crate::Result::Ok(())
         });
     }